+++ include/databento/detail/shared_channel.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>  // size_t
-#include <cstdint>  // uint8_t
+#include <cstddef>  // byte, size_t
 #include <memory>   // shared_ptr
 
 #include "databento/ireadable.hpp"
@@ -13,14 +12,14 @@ class SharedChannel : public IReadable {
   SharedChannel();
 
   // Write `data` of `length` bytes to the channel.
-  void Write(const std::uint8_t* data, std::size_t length);
+  void Write(const std::byte* data, std::size_t length);
   // Signal the end of input.
   void Finish();
   // Read exactly `length` bytes.
-  void ReadExact(std::uint8_t* buffer, std::size_t length) override;
+  void ReadExact(std::byte* buffer, std::size_t length) override;
   // Read at most `length` bytes. Returns the number of bytes read. Will only
   // return 0 if the end of the stream is reached.
-  std::size_t ReadSome(std::uint8_t* buffer, std::size_t length) override;
+  std::size_t ReadSome(std::byte* buffer, std::size_t length) override;
 
  private:
   class Channel;