+++ include/databento/detail/zstd_stream.hpp
@@ -3,7 +3,6 @@
 #include <zstd.h>
 
 #include <cstddef>  // size_t
-#include <cstdint>  // uint8_t
 #include <memory>   // unique_ptr
 #include <vector>
 
@@ -16,19 +15,19 @@ class ZstdDecodeStream : public IReadable {
  public:
   explicit ZstdDecodeStream(std::unique_ptr<IReadable> input);
   ZstdDecodeStream(std::unique_ptr<IReadable> input,
-                   std::vector<std::uint8_t>&& in_buffer);
+                   std::vector<std::byte>&& in_buffer);
 
   // Read exactly `length` bytes into `buffer`.
-  void ReadExact(std::uint8_t* buffer, std::size_t length) override;
+  void ReadExact(std::byte* buffer, std::size_t length) override;
   // Read at most `length` bytes. Returns the number of bytes read. Will only
   // return 0 if the end of the stream is reached.
-  std::size_t ReadSome(std::uint8_t* buffer, std::size_t max_length) override;
+  std::size_t ReadSome(std::byte* buffer, std::size_t max_length) override;
 
  private:
   std::unique_ptr<IReadable> input_;
   std::unique_ptr<ZSTD_DStream, std::size_t (*)(ZSTD_DStream*)> z_dstream_;
   std::size_t read_suggestion_;
-  std::vector<std::uint8_t> in_buffer_;
+  std::vector<std::byte> in_buffer_;
   ZSTD_inBuffer z_in_buffer_;
 };
 
@@ -42,15 +41,15 @@ class ZstdCompressStream : public IWritable {
   ZstdCompressStream& operator=(ZstdCompressStream&&) = delete;
   ~ZstdCompressStream() override;
 
-  void WriteAll(const std::uint8_t* buffer, std::size_t length) override;
+  void WriteAll(const std::byte* buffer, std::size_t length) override;
 
  private:
   ILogReceiver* log_receiver_;
   IWritable* output_;
   std::unique_ptr<ZSTD_CStream, std::size_t (*)(ZSTD_CStream*)> z_cstream_;
-  std::vector<std::uint8_t> in_buffer_;
+  std::vector<std::byte> in_buffer_;
   ZSTD_inBuffer z_in_buffer_;
   std::size_t in_size_;
-  std::vector<std::uint8_t> out_buffer_;
+  std::vector<std::byte> out_buffer_;
 };
 }  // namespace databento::detail