+++ include/databento/file_stream.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>  // size_t
-#include <cstdint>  // uint8_t
+#include <cstddef>  // byte, size_t
 #include <fstream>  // ifstream, ofstream
 #include <string>
 
@@ -14,10 +13,10 @@ class InFileStream : public IReadable {
   explicit InFileStream(const std::string& file_path);
 
   // Read exactly `length` bytes into `buffer`.
-  void ReadExact(std::uint8_t* buffer, std::size_t length) override;
+  void ReadExact(std::byte* buffer, std::size_t length) override;
   // Read at most `length` bytes. Returns the number of bytes read. Will only
   // return 0 if the end of the stream is reached.
-  std::size_t ReadSome(std::uint8_t* buffer, std::size_t max_length) override;
+  std::size_t ReadSome(std::byte* buffer, std::size_t max_length) override;
 
  private:
   std::ifstream stream_;
@@ -27,7 +26,7 @@ class OutFileStream : public IWritable {
  public:
   explicit OutFileStream(const std::string& file_path);
 
-  void WriteAll(const std::uint8_t* buffer, std::size_t length) override;
+  void WriteAll(const std::byte* buffer, std::size_t length) override;
 
  private:
   std::ofstream stream_;