+++ include/databento/ireadable.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>  // size_t
-#include <cstdint>  // uint8_t
+#include <cstddef>  // byte, size_t
 
 namespace databento {
 // An abstract class for readable objects to allow for runtime polymorphism
@@ -11,10 +10,9 @@ class IReadable {
   virtual ~IReadable() = default;
 
   // Read exactly `length` bytes into `buffer`.
-  virtual void ReadExact(std::uint8_t* buffer, std::size_t length) = 0;
+  virtual void ReadExact(std::byte* buffer, std::size_t length) = 0;
   // Read at most `length` bytes. Returns the number of bytes read. Will only
   // return 0 if the end of the stream is reached.
-  virtual std::size_t ReadSome(std::uint8_t* buffer,
-                               std::size_t max_length) = 0;
+  virtual std::size_t ReadSome(std::byte* buffer, std::size_t max_length) = 0;
 };
 }  // namespace databento