+++ tests/src/mock_io.cpp
@@ -6,11 +6,11 @@
 
 using databento::tests::mock::MockIo;
 
-void MockIo::WriteAll(const std::uint8_t* buffer, std::size_t length) {
+void MockIo::WriteAll(const std::byte* buffer, std::size_t length) {
   contents_.insert(contents_.end(), buffer, buffer + length);
 }
 
-void MockIo::ReadExact(std::uint8_t* buffer, std::size_t length) {
+void MockIo::ReadExact(std::byte* buffer, std::size_t length) {
   const auto remaining_bytes = contents_.size() - read_idx_;
   if (remaining_bytes < length) {
     throw std::runtime_error{"Not enough bytes remaining: expected " +
@@ -23,7 +23,7 @@ void MockIo::ReadExact(std::uint8_t* buffer, std::size_t length) {
   read_idx_ += s_length;
 }
 
-std::size_t MockIo::ReadSome(std::uint8_t* buffer, std::size_t max_length) {
+std::size_t MockIo::ReadSome(std::byte* buffer, std::size_t max_length) {
   auto read_size = static_cast<std::ptrdiff_t>(
       std::min(max_length, contents_.size() - read_idx_));
   std::copy(contents_.cbegin() + read_idx_,